+++ packages/xchain-bitcoin/src/index.ts
@@ -1,4 +1,4 @@
 export * from './types'
 export * from './client'
-export { broadcastTx, getDerivePath } from './utils'
+export { broadcastTx, getDerivePath, getDefaultFees, getDefaultFeesWithRates } from './utils'
 export { createTxInfo } from './ledger'